requirements.txt
watchdog
openai-whisper
apscheduler
google-api-python-client
google-auth-oauthlib
openai
pyyaml
